requirements.txt
streamlit
langchain
langchain-google-genai
polars
python-dotenv
faiss-cpu
openpyxl
langgraph
pypdf
python-docx
